Forbid AC bonuses on jewelry
Of Defense, Earth, and Golem are now forbidden on rings and amulets.
Crowns already didn't have these.  I feel it's fair as jewelry isn't
supposed to be as protective as armor, plus it'll make rings that
boost stats and resistances more relevant.  Also, Xeen did that too!

Doom and Earth Affinity are still allowed, as AC there is incidental.