Fix some blaster issues
It seems in MM7 blasters ignore range penalty at any skill instead of at
Expert+, which I'm not okay with as I already put it as the Expert perk
in the description.  So now it's reverted to MM6 rules.  Also, the
Master perk mistakenly used raw skill value (without Squire etc. boni).