Fix two previous commits
In 2e2457c, I did end up breaking the RNG as the saved random seed was
being overwritten by the fixed ones for subsequent NPCs on the map.

For 5dda232, I had to change the location of dismiss reply flag bugfix
hook, as the newly added house hirelings do reset it, but on dialog
start, and my code did it at dialog end, which was too late sometimes.