Hide all join guild topics after paying up
It bothered me a little that after joining a magic guild, the
recruiter's dialog topic disappears, but the other three recruiters
still try to sell membership.  Perhaps it'd be better to change this
the other way, but for now, consistency demands all offers to join
a particular guild must become hidden once membership is obtained.