Make SP drain less random
Vampires and other monsters with the DrainSP special are so annoying in
vanilla that I for one just reload every time they manage to disable my
caster, which is thankfully not too often, as it must pass the level and
resistance checks.  This is bad game design: they might as well crash
the game or force TPK for their resulting impact (namely, the reload)!

So now some SP is drained with each attack that hits, but always just a
fraction of the target's current supply, and stats only offer partial
resistance (four 1/2 rolls like with HP damage).  And since the drain
scales with the current SP pool, I'd say overall it's more merciful now.

I also had to redo the old primary-stat-condition-resistance hooks here.