Improve some more item caches
Building on the previous commit, now elven and dwarven treasuries should
have much better items.  The latter might even house a random artifact
sometimes, although I made sure to avoid the same with elves due to that
room being inaccessible with the light promotion.  The chests in castle
Harmondale, second floor are improved as well, and also lack traps now.
Rationale: the bulter says they're gifts from the party's new allies, so
having them being crappy and trapped was frankly ridiculous story-wise.

It occurs to me that a bunch of early-game locations have guaranteed
and/or random artifact chests in the mod now, unlike in vanilla MM7 (but
similar to MM6 and MM8).  It would've messed up the game for people that
want to collect a specific set of artifacts, if not for the patch
feature that ungenerates unlooted artifacts on refill, and the newly
added refill-triggering hourglasses.  But with these it's probably okay!