Remove floor items from Walls of Mist
I'm 90% sure that part of WoM was initially a different dungeon and NWC
just forgot to remove the item spawners and make it consistent with the
rest of proving grounds.  This also explains the barrels, but I'm
leaving those as I've already nerfed their refill to once per year.

Also here: mentioned the darkness penalty in Torch Light description.