Let travel hirelings boost party move speed
To make them more useful, NPCs that reduce travel times by N days also
increase walking speed by N * 10% of base, either on land or water/lava.
Horsemen only work if you have a Knight horse, and the horses themselves
also grant a land movement bonus equal to their travel time reduction.

With an optimised setup, walking on the ground can actually be almost as
fast as flying now (the latter is not boosted by any of the hirelings).
And of course, the new bonus also works indoors, where you cannot fly!

I've also noticed that in Shoals, all NPCs are hidden, but the party
still gets the benefits of horses and the warlock familiar (the normal
hirelings are disabled, though).  Not sure if this needs fixing.