Alter some monster spells with difficulty
This commit adds a mechanism for dynamically changing monster spell
repertoire depending on current difficulty.  Only the spell ID changes,
not skill, and it's either different on Hard, or on both Normal or Hard.

The idea is to add flavor to some monsters without spoiling the vanilla
experience (which corresponds to Easy).  For now I've added a bunch of
spells to some monsters, including Deadly Swarm which required some
hacks for the game to recognize it, but the meat of it is reserved
for future commits, where I'll code many more new monster spell types.