Add a Regeneration monster spell
It's not a buff, but just a one-time heal, equivalent to the current
Power Cure, but flavored as a natural ability.  As such, trolls and
hydras get it at Normal, and also wraiths and vampires on Hard.

Monsters are progressively less likely to cast it when above 50% HP,
unlike Power Cure.  Later it will differ more as I make the latter AOE.

Also here: added a LOS check to Inferno (although I'm now suspecting
it's unnecessary), and made new spell sounds originate from caster.