Add a ship to Evenmorn before Priest quest
Now the Tatalian pirate island ship is actually interactable, and it can
take you to Evenmorn for 4000 gold base cost.  Instead of completing the
Priest promotion, it requires talking to Lasker -- in the future it will
only work if the party has reached him without killing the thieves.

Kaltenberg has graciously helped again with a pier model for the ship.

Of course the route from Evenmorn back to the island is also added, and
will not work without Lasker either (and the vanilla ship to Tatalia
mainland is restricted by the Priest quest now).  Frankly, paying
several thousand is probably an inferior option to clearing the caves,
but there can be legitimate reasons to delay Priest promotion (such as
maximizing Deyja reputation on the dark path), so I hope it's useful.

The pirate ship got the unused Warlock boat voice because why not.

Also here: sailing to Tatalia and Avlee rotated party the wrong way.