Fix generating MMExt data tables
Just in case someone tries to do it manually: it needs a workaround
to let MMExt recognize that MapStats.txt indices are now off by one.

I actually already had this fix for myself, but I only just realized
I can bundle it with the release.  Compatibility is never for the bad!