Sell lost quest items in the black market
Judge no longer returns lost items; they now appear in the Black Market
for sale.  Accordingly, all tracked items are assigned a nonzero value.

To introduce this feature, an NPC in Tatalia offers a quest to recover a
stolen amulet, which is immediately marked as lost and as such must be
bought from the market; an adivce to avoid combat is also offered.
There's also a rumor (from Judge and others) with the same information.

Moreover, once all PCs are promoted once, the market puts a random
artifact/relic for sale; it will reappear on restock until bought.
Only 10 artifacts are now available as L6 loot, to keep the total at 20.