Add a quest for slotless hirelings
Now the Avlee alchemy teacher also starts a quest that grants a choice
of one of three middle-tier hirelings (squire, apprentice, or scholar)
on completion.  These are treated as quest followers so they don't take
a slot (but they do reduce gold), and can be dismissed temporarily.

The quest itself just requires delving into the School of Sorcery.
I thought the unique reward would be balanced with the lost access
to free scrolls, but in retrospect the player can just regain it with
a hourglass.  Oh well, at least these are not exactly free either.