Add back some penalties to shoplifing
During playtesting it turned out that without mandatory reputation
penalties stealing from shops is way too cheap!  But I still don't want
to bring them back, so let's add "shop-specific thief reputation"
instead, that adds to regional reputation and increases even when you're
not caught (the shopkeeper suspects something).  It's separate per shop
and will be slowly decreased to 0 as you buy and sell there normally.

Also here: if you ARE willing to get caught, the max value of stealable
items is now calculated by doubling the thievery skill (with the +/-2
random modifier) instead of the flat +499.  Rationale: since it hurts
reputation AND shopkeeper wariness AND incurs a fine, it should be more.