Add more autonotes (esp. for wells)
I finally got around to extending the autonote.txt array!  As always, I
had to knock out the MMExt code that was already doing the same (and
which I shamelessly referenced).  For starters, I added an autonote for
one Harmondale fountain that lacked it, as well as enabled a bunch of
others that had text but no gamescript support.  This commit also fixes
an off-by-one error in gamescript autonote checks that in some
circumstances prevented certain well notes from being added to journal.
As well, the EI fire resistance well now stacks with tavern etc. buffs.