Un-nerf necromancers
It has been brought to my attention that the Pit is a joke compared to
vanilla, because both the necromancer (and lich king) projectile attacks
and their Dragon Breath spells are easily resisted compared to vanilla.

Let's try to fix this by making the former physical (instead of magical)
and the latter, fire-physical (instead of fire-poison).  Physical damage
on PCs can only be reduced by armor skills, so this won't be as easy.
On monsters, the new Dragon Breath is not that much stronger, except no
monster is immune to it now (not even medusae).  But it's probably fine.