Print spell element in description
In spellbook and guild popups, the school is printed on the left, but in
MM7 this is redundant as both guilds and spellbook pages are clearly
separated by school.  So let's print resistances attack spells use
instead, since in the mod they're not always obvious.  (Spells that
don't apply damage or debuff just won't have anything in that space.)