Fix reported crash when auto-casting
I haven't reproduced it myself, but from the error log the problem is
I don't initialize the target monster/item for spells that only act on
PCs, but even for those there's some code that checks it, which may
occasionally lead to segfault if I just supply random garbage as the
target.  Now it defaults to 0, which should alleviate the issue.