Redue Arena reward at low levels
There were some complains on the forums that for early parties, Arena
reward money is too generous.  Makes sense, since it's proportional to
PC level, while gold normally obtained by the party is roughly
proportional to XP gain, which grows like at least square of the level.
So the lower the level, the better Arena money seems, comparatively.

I try to fix this by multiplying the reward by level/30 when the level
is below 30, so it also grows like its square.  Arena is still somewhat
exploitable as the hourglasses are sellable for fixed sums of money,
but something could be done to that as well if it will remain a problem.