Reduce the power of monster Rock Blast
Behemoths and trolls can throw rocks at Hard, and they ended up a bit
too brutal because physical damage is hard to resist.  So let's try to
nerf their spell power a bit.  Other monsters with this spell are not
changed, because they're usually less numerous and thus less deadly.