Make monster level matter for pickpocket
In vanilla, monster level was technically added to the difficulty of
stealing from it, but because every other factor was multiplied x100,
only level 100+ monsters were marginally harder to steal from.  This was
stupid, so let's multiply the level x20 for this purpose - overall it
means that 5 monster levels weigh as much as 1 Thievery skill level.
This seems fair, although I might tweak the precise numbers later.