Tweak Altar of Wishes a bit
The third room should be somewhat easier now: the cup is larger, and the
door to portal is taller so that elementals are less likely to get stuck
in it.  Also, in the first room spawn points are now largely in the
alcoves, slightly reducing the immediate pressure on the player.