Enable greetings for extra NPCs in stores
Shops, taverns, temples and other such service houses don't show
greetings when talking with an NPC inside, which makes sense as the
proprietor is not a real NPC anyway.  However, several quest NPCs also
reside in such houses, and they even have pertinent greet lines in
vanilla, which are unfortunately hidden by the same code.  So, let's
limit it to the store's owner, showing the greeting as normal otherwise.