Randomize Ninja quest cipher a bit
To spice things up for players who already know how to enter Master's
tomb (and thus skip raiding the School of Sorcery), there are now three
more possible button sequences (plus the vanilla one), with one selected
at random for the given playthrough.  Naturally, there are also three
more corresponding cipher and code word pairs to go with these.

I have also changed the sounds the buttons make for the correct code to
still generate the old tone sequence - actually, any five button presses
will now make the same four sounds, except the fifth one is changed to
an unused alternative bell chime if the sequence is incorrect.

Granted, a dedicated cheater can just force-decode the cipher without
the code word using an online tool, but hopefully at least some players
will now be tempted to complete the quest as intended by NWC.

And for those who bother to gather the code word but not to sit down
with a pencil, Bill Lasker now deciphers the message for you, provided
you didn't alienate him earlier (i.e. it requires guild membership).

Finally, the scroll with the code word can now only be looted if the
quest is active (which is important to avoid cluttering the black
market), and the tomb can still be opened and entered on Erathia refill.