Raise the price of some specitems
The Perfect Bow and Shadow's Mask got hardcoded properties in this mod,
which roughly elevates them to the level of racial and gender-specific
specitems which all cost 15 grand; Sniper's Quiver also qualifies,
being the bow's counterpart.  Besides, the Cloak of the Sheep is
another incidental quest reward (like these three) and it too is priced
at 15k.  So let's set all their prices to just that.  This fixes #8.