Shift the order button 1px left
Turns out the Russian "close" button is misaligned with the English one
for some reason, and furthermore the widescreen layout is based on the
latter.  I aligned my shop order button to the former, but this led to
it being off in most other game setups, so let's conform to majority.