Let beacon masters respect forbidden spells
Lloyd's Beacon is disabled in a few maps, and for a reason, as setting a
beacon in the Arena or the temple in a bottle is moderately exploitable.
Unfortunately, I forgot that when adding beacon master hirelings, so as
of v4.0 one could e.g. visit Arena every day by setting an NPC beacon
there.  Not great!  So I'm disabling the set beacon hireling function in
those maps now - this also required explicitly forbidding the spell in
Arena: it was hardcoded in vanilla, unlike my gamescript-based solution.
I left the recall beacon function alone, since it doesn't actually break
anything any more than the already allowed Town Portal spell.