Let butler tell you where quest items are
Since we already track quest item location for the black market, let's
add a convenience feature that shares this information with the player,
in case they put an important item in a chest somewhere and forget it.
This is handled by the castle butler, not the judge, because the former
can at least plausibly know where the party has been travelling lately.
Naturally, this is purely informative; unlike vanilla, you won't be
given free copies of quest items, whether they're truly lost or not.

Also here: fix Lasker's Master Dagger tutelage being guild-exclusive
(the script that disables that topic got broken by cab1708 earlier),
and make items left in the proving grounds be properly counted as lost.