Restore entrance animations to late-game temples
For whatever reason, the enterable temples in Celeste and the Pit had
their dungeon entrance animations turned off (but still present in game
files).  It's trivial to restore them, although this necessiates a lot
of clicking to actually proceed, so I also wrote a hack that auto-clicks
the 'enter temple' topic upon selecting the temple entrance NPC.

This commit fixes #9.